Set the river speed to scale with the raft.
The river now flows only slightly faster than the raft is
traveling, and speeds up as the raft speeds up.  This should
make it less disorienting when traveling at high speeds, as
the river will still be keeping up with you and giving you a
stable reference point.

Also fixes a minor graphical bug where the river would flow even
when the game was paused.

Change-Id: Ic0afd348f26c0004497b11968a237c82cb3f936d