Changed the exit flow avoid crash on shutdown.
On OS X, when quitting with command-Q the game would crash because one
thread would end while the other was still running. Rather than quit
immediately when command-Q is entered, the game_exiting_ variable gets
set to true so that the game update thread can see that it should quit.
Without this, the other thread keeps running while the game is shutting
down.

Tested on Linux, OS X and Android.
Bug: 25729226

Change-Id: I84cc6d0c2fa0038bb6b30562830e0b26f1c2be65