Added a landing page for Zooshi Santa Tracker.
Added a relatively simple landing page with pictures to simulate
the buttons for each link.

There is also a static html file that redirects from
`/launch/default/santa` to the doxygen generated file.

Bug: b/25790133

Tested: Tested using generated_docs.py on Ubuntu and viewed
on Chrome. The mobile links were tested on an S5.

Change-Id: I8fb08a3fdde9ae243a48b2bb0a6b75e2cf68f705