Call CreateShadowMap from the correct thread.
Shadow mapping was not updated since animations and multithreading
were added. CreateShadowMap must be called in the render thread,
not the update thread.

Removed Renderer* from RenderPrep() since RenderPrep() is always called
in the update thread, and rendering should happen only in the render
thread.

Tested:
Works on Mac and Linux.

Change-Id: I8c1ad10d199ccc54420ef0a9aec5202782f41b15