Implemented basic shadows.
Added new shaders for static and animated entities, as well as the banks.
There are new shaders for rendering the shadow map on animated entities
as well as for rendering shadows.

In config.json, shadows can be turned on by setting render_shadows to
true.

Tested:
Works on Linux, Nexus 5X, and Nexus 9.

Change-Id: I48d39af06a911dfc42ccfd2f9449574d9b670a73