Render Zooshi score with FlatUI.
First, I removed the old DigitComponent, which manually
rendered textures for each digit of the score.

I added a Render3dTextComponent that uses FlatUI to render
text in 3D world space (on an entity). A Breadboard graph
can be used to update the string rendered by the
Render3dTextComponent.

For entities that wish to have the score rendered on them,
they can use the graph `src/rawassets/graphs/string.json`.
This graph concatenates a string consisting of the current
score, using the new node in `src/modules/ui_string.cpp`,
every frame and updates the Render3dTextData appropriately.

Currently, the font used by FlatUI is the same as the font
used for the menu: `RaviPrakash-Regular.ttf`.

Tested: Tested on Ubuntu and Android.

Change-Id: I93ff6d15f69891758726666a892d01c4d0302bde