Fade shadows close to the shadow map border.
The border of the shadow map is obvious (particularly on rocks) because
shadows slide in and out of view as they enter/exit the shadow map
frustum. The shadows should fade out based on distance from the center
of the shadow map (where the player is positioned).

Bug: Issue 27530394

Tested:
Works on Linux and Nexus 5X.

Change-Id: I0f96be467a087ccdbcad063264f68472bfcb9fb8