Depend on `flatbuffer` lib instead of compiling source.
We used to include the flatbuffer .cpp files directly, but now there
is a `flatbuffer` target in both Android and cmake. Much better to
depend on that since if a new file .cpp file is added to FlatBuffers
the dependent builds won't break.

This fixes a build break in Android caused by missing the
flatbuffers/src/util.cpp file. By happenstance, the cmake build didn't
break because that file was included via other dependent libraries.

Change-Id: Ic0d4005418b0fe20883bb8bf467f6c5102df174c
Tested: Builds and runs on Linux and Android.