Initial support for multiple levels.
Uses a UI element, similar to sushi, to pick which level to use,
which is defined by different entity list files to load.

Moves old level layout into "Endless" which still wraps around,
while adding a new level "Easy" that is currently a copy of
Endless, but with a rail that goes in a straight line.

Allows levels to wrap as before, or finish at a specified point along
the rail, used for levels meant to be played through once.

Tested on Android and Linux (by removing Firebase temporarily)

Change-Id: Ib0a09a8128590020f7415b1b230164480d404f04