Adds a system to track player xp, for unlockables.
Adds an XP system, which is used to track how a player is doing.
This is hooked into the unlockable system to give rewards.

Bonuses are tracked which can be applied to adjust the base xp,
and will be hooked into Firebase in a future change.

Tested on Android and OSX.

Change-Id: Ia6a143c597fc8b05742e061421c53d958879f82a