Initial work for Firebase AdMob, Rewarded Video
Adds support for using AdMob rewarded video, by offering the
option to watch a video from the main screen. Doing so grants a
bonus that is applied on the next play session.

This does not handle Pause/Resume when leaving/entering app, as
SDL is making that a bit tricky, and so will be handled in a follow
up change.

Also reorganizes the Firebase helpers/listeners under Game, with
World having pointers to them.

Change-Id: I12446626fe23fc84387f673fa3b9ac59eb75e552